requirements.txt
flask==3.0.3
gunicorn==23.0.0
openai==1.30.1
httpx==0.27.0
flask-cors==4.0.1
pinecone[grpc]==7.3.0
python-dotenv==1.0.1 